python/requirements.txt
# SAQ Python dependencies
# Install with: pip install -r requirements.txt

numpy>=1.20.0
faiss-cpu>=1.7.0

# For GPU support, replace faiss-cpu with:
# faiss-gpu>=1.7.0
